+++ app.js
@@ -14,6 +14,7 @@ require('./config/passport');
 const indexRouter = require("./routes/index");
 const usersRouter = require("./routes/users");
 const productsRouter = require("./routes/products");
+const shopRouter = require("./routes/shop");
 
 var app = express();
 
@@ -44,6 +45,7 @@ app.use(function (req, res, next) {
 app.use("/", indexRouter);
 app.use("/users", usersRouter);
 app.use("/products", productsRouter);
+app.use('/shop', shopRouter);
 
 // catch 404 and forward to error handler
 app.use(function (req, res, next) {
