+++ config/passport.js
@@ -19,6 +19,7 @@ passport.use(
           googleId: profile.id,
           email: profile.emails[0].value,
           avatar: profile.photos[0].value,
+          username: profile.emails[0].value.split("@")[0],
         });
         return cb(null, user);
       } catch (err) {
