+++ public/stylesheets/style.css
@@ -10,7 +10,16 @@
 }
 body {
   background-color: rgba(255, 255, 255, 0.9);
+  margin: 0;
+  height: 100vh;
+  display: grid;
+  grid-template-rows: 50px 1fr 50px;
+  grid-template-areas: 
+    "nav"
+    "main"
+    "footer"
 }
+
 nav {
   display: flex;
   align-items: center;
@@ -24,6 +33,7 @@ nav {
   width: 100%;
   z-index: 1;
   background-color: rgba(0, 0, 0, 0.9);
+  grid-area: nav;
 }
 
 .logo {
@@ -69,12 +79,13 @@ nav ul li a:hover {
 }
 
 footer {
-  position: absolute;
+  position: fixed;
   right: 0;
-  /* bottom: 0; */
   left: 0;
   background: #111;
   color: #fff;
+  bottom: 0;
+  grid-area: footer;
 }
 
 .socials {
@@ -237,4 +248,12 @@ footer {
   text-align: center;
 }
 
-/* temporary --> */
+.container {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 20px;
+  overflow-y: scroll;
+  margin-top: 40px;
+  grid-area: main;
+}
