+++ routes/users.js
@@ -8,7 +8,5 @@ router.get('/:userId/items', usersCtrl.usrItmPg);
 router.post('/:userId/items/add', usersCtrl.addItm);
 router.post('/:userId/removeItems', usersCtrl.rmvItm);
 router.post('/:userId/updateUsername', usersCtrl.updtUsrNm)
-// GET /users/:username - sellers shop page
-router.get("/:username", usersCtrl.shop);
 
 module.exports = router;
