+++ views/partials/header.ejs
@@ -15,22 +15,22 @@
       referrerpolicy="no-referrer"
     />
   </head>
-  <body>
-    <nav>
-      <h2 class="logo">
-        <a href="/???"><span>GA</span>Marketplace</a>
-      </h2>
-      <ul>
-        <li><a href="/products/">All Products</a></li>
-        <li><a href="/???">???</a></li>
-        <li><a href="/???">???</a></li>
-        <li><a href="/???">???</a></li>
-      </ul>
+  
+  <nav>
+    <h2 class="logo">
+      <a href="/???"><span>GA</span>Marketplace</a>
+    </h2>
+    <ul>
+      <li><a href="/products/">All Products</a></li>
+      <li><a href="/???">???</a></li>
+      <li><a href="/???">???</a></li>
+      <li><a href="/???">???</a></li>
+    </ul>
 
-      <% if (user) { %>
-      <a href="/???" class="btn">Profile</a>
-      <a href="/logout" class="btn">Log Out</a>
-      <% } else { %>
-      <a href="/auth/google" class="btn">Log In</a>
-      <% } %>
-    </nav>
+    <% if (user) { %>
+    <a href="/users/<%= user._id %>" class="btn">Profile</a>
+    <a href="/logout" class="btn">Log Out</a>
+    <% } else { %>
+    <a href="/auth/google" class="btn">Log In</a>
+    <% } %>
+  </nav>
